pyproject.toml
[tool.poetry]
name = "todoApi"
version = "0.1.0"
description = ""
authors = ["ghous-mohammad <ghosmohammad247@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
